requirements.txt
aiofiles==24.1.0
annotated-types==0.6.0
anyio==4.3.0
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
docxcompose==1.4.0
et-xmlfile==1.1.0
fastapi==0.110.1
Flask==3.0.3
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
reportlab==4.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.29.0
waitress==3.0.0
Werkzeug==3.0.2
